requirements.txt
discord.py-self
colorama
curl_cffi
httpx
python-dotenv
pyYAML
requests
openai
fastapi
uvicorn
jinja2
python-multipart
itsdangerous